Will update the priority label and moves the model in model folder.
